create separate StarItem component
