+++ src/components/ui/grade/grade.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import s from '../grade/grade.module.scss'
 
-import { Star, StarOutline } from '@/assets/icons'
+import { StarItem } from './star'
 
 export type RatingValue = 0 | 1 | 2 | 3 | 4 | 5
 
@@ -20,24 +20,5 @@ export const Grade: FC<GradeProps> = ({ value, onClick }) => {
       <StarItem value={4} onClick={onClick} selected={value > 3} />
       <StarItem value={5} onClick={onClick} selected={value > 4} />
     </div>
-  )
-}
-
-type StarItemProps = {
-  value: RatingValue
-  selected: boolean
-  onClick: (value: RatingValue) => void
-}
-
-const StarItem: FC<StarItemProps> = ({ value, selected, onClick }) => {
-  return (
-    <button
-      onClick={() => {
-        onClick(value)
-      }}
-      className={s.button}
-    >
-      {selected ? <Star /> : <StarOutline />}
-    </button>
   )
 }
