+++ src/components/ui/grade/star/index.ts
@@ -1,0 +1,1 @@
+export * from './star'
